+++ scripts/subnet/setup.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
+ROOT=${1:-$HOME}
+
 # Go to the repository
-cd $HOME/Subnet_S
+cd $ROOT/SubVortex
+
+# Install speedtest-cli (needed only for the miner)
+apt-get update
+apt-get install speedtest-cli 
 
 # Install the dependencies
 python -m pip install -e .