+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 
+
+## 
+
 ## 1.0.1 / 2024-02-23
 
 **Full Changelog**: https://github.com/eclipsevortex/SubVortex/commits/v1.0.1