+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-## 1.0.1 / 2024-02-23
+## 0.1.0 / 2024-02-24
 
-**Full Changelog**: https://github.com/eclipsevortex/SubVortex/commits/v1.0.1
+## What's Changed
+* feat: implement the release scripts by @eclipsevortex in https://github.com/eclipsevortex/SubVortex/pull/1
 
+## New Contributors
+* @eclipsevortex made their first contribution in https://github.com/eclipsevortex/SubVortex/pull/1
 
+**Full Changelog**: https://github.com/eclipsevortex/SubVortex/commits/v0.1.0